Restore missing GUI options and metrics
